deployment/requirements.txt
streamlit
Pillow
numpy
tensorflow
transformers